requirements-mac.txt
certifi==2018.4.16
chardet==3.0.4
click==6.7
Flask==1.0.2
idna==2.7
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
netifaces==0.10.7
pywebview==2.0.3
requests==2.19.1
urllib3==1.23
Werkzeug==0.14.1
pyobjc
PyQt5
